refactored substanced scaffold to have a separate 'retail' package; moved templates into that; removed retails viwews from the original views.py
