+++ substanced/scaffolds/substanced/+package+/retail/templates/splash.pt
